+++ src/components/Header/Header.tsx
@@ -4,12 +4,13 @@ import * as React from 'react';
 import { PhaseMenu } from '../PhaseMenu';
 import { UserList } from '../UserList';
 import { UserMenu } from '../UserMenu';
-import { BoardUsers, RetroMode } from '../../types';
+import { BoardUsers } from '../../types';
 import Logo from './subcomponents/Logo';
 
 import './Header.css';
 import { mapStateToProps } from './Header.container';
 import { connect } from 'react-redux';
+import { RetroMode } from '../../constants/mode';
 
 export interface OwnHeaderProps {
   boardId: string;
