+++ src/constants/Retrospective.ts
@@ -1,6 +1,6 @@
 import { AddCardTheme } from '../components/AddCard';
 import { IconNames } from '../components/Icon';
-import { RetroMode } from '../types';
+import retroModes, { RetroMode } from './mode';
 
 export type ColumnType = 'positive' | 'negative' | 'actions';
 
@@ -38,402 +38,6 @@ export interface PhaseConfiguration {
 export interface IndexedPhaseConfiguration extends PhaseConfiguration {
   index: number;
 }
-
-const leanCoffeeModeConfiguration: PhaseConfiguration[] = [
-  {
-    name: 'Lean Coffee',
-    description: 'Create cards, communicate your thoughts & stack common',
-    columns: [
-      {
-        id: 'positive',
-        name: 'Lean Coffee',
-        type: 'positive',
-        voting: {
-          enabled: false,
-          displayed: false
-        },
-        focus: {
-          enabled: true,
-          column: 'positive',
-          align: 'right'
-        },
-        sorted: false
-      }
-    ],
-    activities: [
-      { icon: 'phase1-create', description: 'Create cards' },
-      { icon: 'phase1-communicate', description: 'Communicate meaning' },
-      { icon: 'phase1-combine', description: 'Combine to stack' }
-    ]
-  },
-  {
-    name: 'Vote',
-    description: 'Vote on the most important cards for yourself',
-    columns: [
-      {
-        id: 'positive',
-        name: 'Lean Coffee',
-        type: 'positive',
-        voting: {
-          enabled: true,
-          displayed: false
-        },
-        focus: {
-          enabled: false,
-          column: 'positive',
-          align: 'right'
-        },
-        sorted: false
-      }
-    ],
-    activities: [{ icon: 'phase2-vote', description: 'Vote wisely' }]
-  },
-  {
-    name: 'Discuss',
-    description: 'Discuss top-voted cards',
-    columns: [
-      {
-        id: 'positive',
-        name: 'Lean Coffee',
-        type: 'positive',
-        voting: {
-          enabled: false,
-          displayed: true
-        },
-        focus: {
-          enabled: true,
-          column: 'positive',
-          align: 'right'
-        },
-        sorted: false
-      }
-    ],
-    activities: [{ icon: 'phase3-discuss', description: 'Discuss top-voted' }]
-  }
-];
-
-const simpleRetroModeConfiguration: PhaseConfiguration[] = [
-  {
-    name: 'Write',
-    description: 'Create cards, communicate your thoughts & stack common',
-    columns: [
-      {
-        id: 'positive',
-        name: 'Positive',
-        type: 'positive',
-        voting: {
-          enabled: false,
-          displayed: false
-        },
-        focus: {
-          enabled: true,
-          column: 'positive',
-          align: 'right'
-        },
-        sorted: false
-      },
-      {
-        id: 'negative',
-        name: 'Negative',
-        type: 'negative',
-        voting: {
-          enabled: false,
-          displayed: false
-        },
-        focus: {
-          enabled: true,
-          column: 'negative',
-          align: 'left'
-        },
-        sorted: false
-      }
-    ],
-    activities: [
-      { icon: 'phase1-create', description: 'Create cards' },
-      { icon: 'phase1-communicate', description: 'Communicate meaning' },
-      { icon: 'phase1-combine', description: 'Combine to stack' }
-    ]
-  },
-  {
-    name: 'Vote',
-    description: 'Vote on the most important cards for yourself',
-    columns: [
-      {
-        id: 'positive',
-        name: 'Positive',
-        type: 'positive',
-        voting: {
-          enabled: true,
-          displayed: false
-        },
-        focus: {
-          enabled: false,
-          column: 'positive',
-          align: 'right'
-        },
-        sorted: false
-      },
-      {
-        id: 'negative',
-        name: 'Negative',
-        type: 'negative',
-        voting: {
-          enabled: true,
-          displayed: false
-        },
-        focus: {
-          enabled: false,
-          column: 'negative',
-          align: 'left'
-        },
-        sorted: false
-      }
-    ],
-    activities: [{ icon: 'phase2-vote', description: 'Vote wisely' }]
-  },
-  {
-    name: 'Discuss',
-    description: 'Discuss top-voted cards and define actions',
-    columns: [
-      {
-        id: 'positive',
-        name: 'Positive',
-        type: 'positive',
-        voting: {
-          enabled: false,
-          displayed: true
-        },
-        focus: {
-          enabled: true,
-          column: 'actions',
-          align: 'right'
-        },
-        sorted: false
-      },
-      {
-        id: 'negative',
-        name: 'Negative',
-        type: 'negative',
-        voting: {
-          enabled: false,
-          displayed: true
-        },
-        focus: {
-          enabled: true,
-          column: 'actions',
-          align: 'right'
-        },
-        sorted: false
-      },
-      {
-        id: 'actions',
-        name: 'Actions',
-        type: 'actions',
-        voting: {
-          enabled: false,
-          displayed: true
-        },
-        focus: {
-          enabled: false,
-          column: 'actions',
-          align: 'right'
-        },
-        sorted: true
-      }
-    ],
-    activities: [
-      { icon: 'phase3-discuss', description: 'Discuss top-voted' },
-      { icon: 'phase3-define', description: 'Define actions' }
-    ]
-  }
-];
-
-const startStopContinueRetroModeConfiguration: PhaseConfiguration[] = [
-  {
-    name: 'Write',
-    description: 'Create cards, communicate your thoughts & stack common',
-    columns: [
-      {
-        id: 'start',
-        name: 'Start',
-        type: 'positive',
-        voting: {
-          enabled: false,
-          displayed: false
-        },
-        focus: {
-          enabled: true,
-          column: 'start',
-          align: 'right'
-        },
-        sorted: false
-      },
-      {
-        id: 'stop',
-        name: 'Stop',
-        type: 'negative',
-        voting: {
-          enabled: false,
-          displayed: false
-        },
-        focus: {
-          enabled: true,
-          column: 'stop',
-          align: 'left'
-        },
-        sorted: false
-      },
-      {
-        id: 'continue',
-        name: 'Continue',
-        type: 'positive',
-        voting: {
-          enabled: false,
-          displayed: false
-        },
-        focus: {
-          enabled: true,
-          column: 'continue',
-          align: 'right'
-        },
-        sorted: false
-      }
-    ],
-    activities: [
-      { icon: 'phase1-create', description: 'Create cards' },
-      { icon: 'phase1-communicate', description: 'Communicate meaning' },
-      { icon: 'phase1-combine', description: 'Combine to stack' }
-    ]
-  },
-  {
-    name: 'Vote',
-    description: 'Vote on the most important cards for yourself',
-    columns: [
-      {
-        id: 'start',
-        name: 'Start',
-        type: 'positive',
-        voting: {
-          enabled: true,
-          displayed: false
-        },
-        focus: {
-          enabled: false,
-          column: 'start',
-          align: 'right'
-        },
-        sorted: false
-      },
-      {
-        id: 'stop',
-        name: 'Stop',
-        type: 'negative',
-        voting: {
-          enabled: true,
-          displayed: false
-        },
-        focus: {
-          enabled: false,
-          column: 'stop',
-          align: 'left'
-        },
-        sorted: false
-      },
-      {
-        id: 'continue',
-        name: 'Continue',
-        type: 'positive',
-        voting: {
-          enabled: true,
-          displayed: false
-        },
-        focus: {
-          enabled: false,
-          column: 'continue',
-          align: 'right'
-        },
-        sorted: false
-      }
-    ],
-    activities: [{ icon: 'phase2-vote', description: 'Vote wisely' }]
-  },
-  {
-    name: 'Discuss',
-    description: 'Discuss top-voted cards and define actions',
-    columns: [
-      {
-        id: 'start',
-        name: 'Start',
-        type: 'positive',
-        voting: {
-          enabled: false,
-          displayed: true
-        },
-        focus: {
-          enabled: true,
-          column: 'actions',
-          align: 'right'
-        },
-        sorted: false
-      },
-      {
-        id: 'stop',
-        name: 'Stop',
-        type: 'negative',
-        voting: {
-          enabled: false,
-          displayed: true
-        },
-        focus: {
-          enabled: true,
-          column: 'actions',
-          align: 'right'
-        },
-        sorted: false
-      },
-      {
-        id: 'continue',
-        name: 'Continue',
-        type: 'positive',
-        voting: {
-          enabled: false,
-          displayed: true
-        },
-        focus: {
-          enabled: true,
-          column: 'actions',
-          align: 'right'
-        },
-        sorted: false
-      },
-      {
-        id: 'actions',
-        name: 'Actions',
-        type: 'actions',
-        voting: {
-          enabled: false,
-          displayed: true
-        },
-        focus: {
-          enabled: false,
-          column: 'actions',
-          align: 'right'
-        },
-        sorted: true
-      }
-    ],
-    activities: [
-      { icon: 'phase3-discuss', description: 'Discuss top-voted' },
-      { icon: 'phase3-define', description: 'Define actions' }
-    ]
-  }
-];
-
-const retroModes = {
-  lean: leanCoffeeModeConfiguration,
-  positiveNegative: simpleRetroModeConfiguration,
-  startStopContinue: startStopContinueRetroModeConfiguration
-};
 
 export function getPhasesCount(retroMode: RetroMode) {
   return retroModes[retroMode].length;
